[Refactor] DeleteHandler used.
