--- server/deleteHandlers
@@ -1,10 +1,0 @@
-var logger = require('bristol');
-var db = require('./db');
-var utils = require('./lib/utils');
-
-module.exports = {
-  //delete route function for events
-  eventDEL: function() {
-
-  }
-}
